codes/requirements.txt
absl-py==0.15.0
attrs==21.4.0
Bottleneck==1.3.4
brotlipy==0.7.0
captum==0.5.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.4
cryptography==37.0.1
cycler==0.11.0
Cython==0.29.32
docopt==0.6.2
et-xmlfile==1.1.0
fonttools==4.25.0
grpcio==1.42.0
hdbscan==0.8.28
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
joblib==1.1.0
kaleido==0.2.1
kiwisolver==1.4.2
Markdown==3.3.4
matplotlib==3.5.1
munkres==1.1.4
networkx==2.8.6
numexpr==2.8.1
numpy==1.22.3
opencv-python
opencv-python-headless
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
patsy==0.5.2
Pillow==9.0.1
pip==21.2.4
pipreqs==0.4.11
plotly==5.6.0
pluggy==1.0.0
prettytable==3.3.0
protobuf==3.20.1
py==1.11.0
pycparser==2.21
PyKrige==1.7.0
pyOpenSSL==22.0.0
pyparsing==3.0.4
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
setuptools==61.2.0
sip==6.5.1
six==1.16.0
statsmodels==0.13.2
tenacity==8.0.1
tensorboard==2.0.0
threadpoolctl==2.2.0
toml==0.10.2
tomli==2.0.1
torch==1.11.0
torchaudio==0.11.0
torchmetrics==0.9.3
torchvision==0.12.0
tornado==6.1
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.0.3
wheel==0.37.1
yarg==0.1.9
zipp==3.8.0
